Re-indent the file to 4 spaces, ctermbg=NONE
The 2 spaces were harder to read.

The ctermbg=NONE setting was being triggered every time the colorscheme
changed in order to keep the terminal background transparent. When
combined with Denite, airline, and base16, however, it caused the
statusline to flicker whenever a Denite buffer exited, but only if a
base16 colorscheme was used. I don't know yet which plugin is at fault.
Therefore I've temporarily disabled the ctermbg=NONE setting.