Replace lspinstall with lsp-install-server
Required much less code, which is nice. Also cleaned up the on_attach
function a little.

Removing the arduino-language-server config for now. Not currently using
it, plus there is now a config included with lspconfig. There is an
issue currently open with lsp-install-server to get it added there too.
Hopefully by the time I want it again it'll all be sorted.